requirements.txt
alembic==1.4.2
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
